:tada: emoji dropdown works again
